requirements.txt
Flask
line-bot-sdk==3.17.1
gunicorn
waitress
python-dotenv
requests
supabase
openai
psycopg2-binary
pandas
openpyxl
Pillow